requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pymysql==1.1.1
python-dotenv==1.0.1
weasyprint==68.1
jinja2==3.1.4
